refactor folder and names
