+++ src/components/atoms/CloseButton/CloseButton.scss
@@ -1,6 +1,6 @@
 @import "colors";
 
-.CloseModal {
+.CloseButton {
   position: absolute;
   right: 15px;
   top: 15px;
