+++ src/components/modals/ModalDeleteToken/ModalDeleteToken.scss
