+++ src/components/molecules/Modal/Modal.tsx
@@ -1,7 +1,7 @@
+import { CloseButton } from 'components/atoms/CloseButton/CloseButton'
+import { Overlay } from 'components/atoms/Overlay/Overlay'
 import * as React from 'react'
-import { CloseModal } from './CloseModal'
 import './Modal.scss'
-import { Overlay } from './Overlay'
 
 interface ModalProps {
   readonly onClose: () => void
@@ -12,7 +12,7 @@ interface ModalProps {
 export const Modal = (props: ModalProps) => (
   <Overlay show={props.show}>
     <div className={'Modal'}>
-      <CloseModal onClose={props.onClose} />
+      <CloseButton onClose={props.onClose} />
       {props.children}
     </div>
   </Overlay>
